+++ src/bda/plone/orders/browser/bookings.py
@@ -6,7 +6,6 @@ from bda.plone.orders import interfaces as ifaces
 from bda.plone.orders import message_factory as _
 from bda.plone.orders import permissions
 from bda.plone.orders import vocabularies as vocabs
-from bda.plone.orders.browser.common import ContentViewBase
 from bda.plone.orders.browser.common import customers_form_vocab
 from bda.plone.orders.browser.common import salaried_form_vocab
 from bda.plone.orders.browser.common import states_form_vocab
@@ -128,7 +127,7 @@ class BookingSalariedTransition(BookingTransition):
 # bookings views
 ###############################################################################
 
-class BookingsView(ContentViewBase):
+class BookingsView(BrowserView):
     table_view_name = '@@bookingstable'
 
     def __init__(self, context, request):